requirements.txt
torch
torchvision
stanza
spacy-transformers